Simplify range of lv2:designation
The previous range was not incorrect, but suggests inferring the type of
objects by the property range.  In LV2 (and sord_validate), we take a
stricter stance, and use range to check that object types are explicitly
of a valid type, so aggregate ranges are problematic.

Since lv2:Designation is an rdfs:subClassOf rdf:Property anyway, it's
easier to just set the range to rdf:Property.