Remove domain of lv2:scalePoint
This property is useful for both control ports and properties, so the
previous domain caused validation errors.  Since there is no simple
class which covers both ports and properties, and this property is
probably useful in other cases as well, simply removing the domain makes
sense.

Note that having multiple domains for a property implies that any
resource with that property has all the domain types (not just one), so
adding lv2:Parameter as a domain as well would be incorrect.