Remove unnecessary struct names
These are not necessary since they are not recursive types, and the names
violated the C standard.  There also seems to be a bug in Doxygen where
TYPEDEF_HIDES_STRUCT was not working, which made the documentation confusing
and ugly.

Strictly speaking, this could break user code that used these names directly,
but this was never intended to be supported and should be fixed if it actually
occurs anywhere.