Use the same name for recursive struct types
Similar to the previous commit, this was non-standard and caused documentation
issues.  With this commit, the data structures index page looks as expected.

Unlike the previous commit, though, these names are actually needed because the
struct definition refers to itself.  Simply use the same name instead of the
underscore prefix, since the "struct namespace" in C is different anyway.