Fix state:StateChanged URI in metadata and documentation
This was a mistake, and it's arguable which URI was correct before.  It was
supposed to be "StateChanged" since the name of URIs should stand
alone (ignoring the conventional prefix, even though this has been screwed up
in several places in the past).  Normally, the specification in Turtle is
considered canonical for URIs, but in this case, the URI defined in the
corresponding header used "StateChanged" (both in name and value), as did all
the news.  Really everything except the subject in the metadata.

So, consider this a documentation bug, and fix the URI in the
spec/documentation to match the implemented one.