Modernize compiler checks in attributes.h
These caused warnings in compilers where __GNUC__ is not defined.  That could
be fixed, but it makes things quite ugly, these versions are quite old at this
point, and these features are just for developers.  So, instead, simply check
the major version instead of the specific minor version to keep things terse.