Remove unused datatypes from xsd.ttl
This data costs memory, and I think it is better to not present things that
aren't useful to LV2 implementations at all here.

The whitespace annotations could be useful for cleaning up literals, but this
isn't currently used (and tends to be baked in to implementations anyway), so
remove them as well.  If necessary they can be added later easily enough.