Use matching datatypes for fixed-size bounds in xsd.ttl
This avoids the need to do any datatype conversion for handling bounds.  In
particular, it avoids troublesome range issues, like the maximum
xsd:unsignedLong being too large to fit in an xsd:long, which is likely what an
implementation without bigint support uses to represent xsd:integer.

So, avoid all of that by simply using the exact datatype in bounds for that
datatype, which makes more sense anyway.