midigate: Fix output timing
Previously, the state was updated before writing the output (up to the current
time), so output "in the past" would be incorrectly written based on the
current state.  This fixes that by switching things around, so the output is
first written up to the current time, then the state is updated (and the loop
continues).

This takes advantage of the fact that write_output() effectively does nothing
with a zero length.