Add missing release metadata
These should have been added for the 1.18.4 release.  This seems a bit silly
for minor changes like this that span the entire project, but the documentation
tools currently require it.

Simplifying this would be nice, perhaps by using NEWS files as the canonical
source instead, which would be a nicer and more typical developer experience.