lv2specgen: Remove warnings on external terms
This conveniently pointed out potential issues in the LV2 documentation, but
makes the tool very noisy to use on anything else.  Referring to some external
thing really shouldn't be treated like a warning in linked data.  Preventing
mistakes (where full URIs are in the output instead of the expected prefixed
names) will have to be done in some other, better way (probably based on an
explicitly given set of prefixes that should be documented).