Use standard ISC license text
I'm not sure where the disclaimer variant that starts with "THIS" came
from (although it is better that way), but this one that starts with "THE" is
the standard text which is recognized by tools like `licensee` (and therefore
Github) as a perfect match.  There are no other changes other than whitespace.