Remove double size check entirely
It turns out that, since this header is widely included in dependants, removing
the conditional in the previous commit triggers undocumented warnings
downstream.

Since the check is pretty esoteric anyway, and has been widely included across
most of the LV2 ecosystem for a decade and never once come up, I think it's
safe to just delete this and be done with the hassle.