Use bare boolean literals
I'm not sure why these were written out in full, although vaguely recall
something about rdflib having trouble with these.  In any case, it seems to
work now, and this is fundamental Turtle syntax that should be supported by
everything.

This is necessary for the lint tests to pass after improvements to serd that
will rewrite these literals in this way.