Suppress GCC Wstringop-overflow warning in atom overflow test
In release builds, GCC 14.2.1 reports "error: writing 4 bytes into a region of
size between 18446744073709551581 and 3".

I'm pretty sure this is a compiler bug: the message suggests an underflow, the
flagged code does have an overflow check (being the whole point of the test),
and the test passes in both ASan and Valgrind.