Remove static constant for clang nonblocking inference
Unfortunately the function effect checking added in clang 20 won't infer a
function is nonblocking if it has a static constant variable (because the
compiler might insert a lock to initialize it, or it might have an advanced
constructor in C++). This isn't a big loss here, so just inline the value.