Avoid checking for tools and options that won't be used
With many components disabled, configuration checked for many programs and
compiler flags that weren't actually used for anything.  This is slow, and
makes simple builds look a lot more complicated than they are.  So, rework
things to only check for tools and flags that will actually be used.

Mainly this is to make the package nice to use as a subproject.